Move specialization in order to avoid cyclic dependency problem
